requirements.txt
librosa==0.8.1
opencv-python==4.6.0.66
tqdm==4.45.0
numba==0.48
numpy==1.23.0
Pillow==9.5.0
scipy==1.10.1

sdk-obs-python==3.23.5
uvicorn[standard]
fastapi
requests
pyyaml


dlib==19.17.0
opencv-python==4.6.0.66

mediapipe==0.10.1

yapf==0.40.1
lmdb==1.4.1

moviepy==1.0.3
pydub==0.25.1